Work in progress: duplicating termemu.cc example
This is a WIP conversion of the termemu.cc mosh example into a
Golang io.ReadWriteCloser interposer. Once completed, it should be
capable of interposing on a terminal session, sending coherent
deltas to the client, and be able to generate patches that take
raw terminals to the current state of the interposed terminal.

This example does not make use of prediction, but is a stepping
stone to that functionality.