WIP predictor: several methods expect epoch (not real) time
Temporary workaround in this code, but it needs to be fixed in
the go-mosh library. Prediction response is much less blinky/
flickery now, but blinking still happens sometimes, the cursor
position still ends up in the wrong place frequently, and some
issues when new lines are added to the bottom of the screen.