New ring delayer implementation
Delays to the ring delayer are scheduled based on the time the write
is received, rather than just injecting a sleep on the write call. This
provides a closer approximation of network delay. However, the delay is
only applied to writes. Reads are instantaneous. Should be sufficient
for latency simulation tests, even if it isn't totally realistic.