Started cleanup of TODOs in interposer
Also added some notes about the predictor's SetSendInterval purpose.
This seems to directly drive whether or not underlines are drawn for
predictions. This needs to be adjusted (with epoch latency?) so that
the low latency typical mode doesn't result in constant flickering
prediction underlines that are resolved within a couple centiseconds.